+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -64,8 +64,8 @@ class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
 
 	// MARK: - Helpers
 		
-	private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, loader: LoaderSpy) {
-		let loader = LoaderSpy()
+	private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, loader: FeedImageDataLoaderSpy) {
+		let loader = FeedImageDataLoaderSpy()
 		let sut = FeedImageDataLoaderCacheDecorator(decoratee: loader)
 		trackForMemoryLeaks(loader, file: file, line: line)
 		trackForMemoryLeaks(sut, file: file, line: line)
@@ -95,33 +95,4 @@ class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
 		wait(for: [exp], timeout: 1.0)
 	}
 	
-	private class LoaderSpy: FeedImageDataLoader {
-		private var messages = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-
-		private(set) var cancelledURLs = [URL]()
-
-		var loadedURLs: [URL] {
-			return messages.map { $0.url }
-		}
-
-		private struct Task: FeedImageDataLoaderTask {
-			let callback: () -> Void
-			func cancel() { callback() }
-		}
-
-		func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-			messages.append((url, completion))
-			return Task { [weak self] in
-				self?.cancelledURLs.append(url)
-			}
-		}
-		
-		func complete(with error: Error, at index: Int = 0) {
-			messages[index].completion(.failure(error))
-		}
-		
-		func complete(with data: Data, at index: Int = 0) {
-			messages[index].completion(.success(data))
-		}
-	}
 }
