+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderWithFallbackCompositeTests.swift
@@ -90,9 +90,9 @@ class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase {
 
 	// MARK: - Helpers
 	
-	private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, primary: LoaderSpy, fallback: LoaderSpy) {
-		let primaryLoader = LoaderSpy()
-		let fallbackLoader = LoaderSpy()
+	private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, primary: FeedImageDataLoaderSpy, fallback: FeedImageDataLoaderSpy) {
+		let primaryLoader = FeedImageDataLoaderSpy()
+		let fallbackLoader = FeedImageDataLoaderSpy()
 		let sut = FeedImageDataLoaderWithFallbackComposite(primary: primaryLoader, fallback: fallbackLoader)
 		trackForMemoryLeaks(primaryLoader, file: file, line: line)
 		trackForMemoryLeaks(fallbackLoader, file: file, line: line)
@@ -122,35 +122,5 @@ class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase {
 		
 		wait(for: [exp], timeout: 1.0)
 	}
-	
-	private class LoaderSpy: FeedImageDataLoader {
-		private var messages = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-
-		private(set) var cancelledURLs = [URL]()
-
-		var loadedURLs: [URL] {
-			return messages.map { $0.url }
-		}
-
-		private struct Task: FeedImageDataLoaderTask {
-			let callback: () -> Void
-			func cancel() { callback() }
-		}
-
-		func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-			messages.append((url, completion))
-			return Task { [weak self] in
-				self?.cancelledURLs.append(url)
-			}
-		}
 		
-		func complete(with error: Error, at index: Int = 0) {
-			messages[index].completion(.failure(error))
-		}
-		
-		func complete(with data: Data, at index: Int = 0) {
-			messages[index].completion(.success(data))
-		}
-	}
-	
 }
